Refactor: last files and functions restructuration
